# Some optimisations for reducing number of memory allocations and improving BVH build speed.