--- a/Jolt/Physics/Collision/Shape/MeshShape.cpp
+++ b/Jolt/Physics/Collision/Shape/MeshShape.cpp
@@ -92,8 +92,8 @@ MeshShapeSettings::MeshShapeSettings(VertexList inVertices, IndexedTriangleList 
 void MeshShapeSettings::Sanitize()
 {
 	// Remove degenerate and duplicate triangles
-	UnorderedSet<IndexedTriangle> triangles;
-	triangles.reserve(mIndexedTriangles.size());
+	UnorderedSet<IndexedTriangle> triangles(/*empty_key=*/IndexedTriangle(std::numeric_limits<uint32>::max(), std::numeric_limits<uint32>::max(), std::numeric_limits<uint32>::max()),
+		/*expected num items=*/mIndexedTriangles.size());
 	TriangleCodec::ValidationContext validation_ctx(mIndexedTriangles, mTriangleVertices);
 	for (int t = (int)mIndexedTriangles.size() - 1; t >= 0; --t)
 	{
@@ -261,8 +261,7 @@ void MeshShape::sFindActiveEdges(const MeshShapeSettings &inSettings, IndexedTri
 
 	// Build a list of edge to triangles
 	using EdgeToTriangle = UnorderedMap<Edge, TriangleIndices, EdgeHash>;
-	EdgeToTriangle edge_to_triangle;
-	edge_to_triangle.reserve(ioIndices.size() * 3);
+	EdgeToTriangle edge_to_triangle(/*empty key=*/Edge(std::numeric_limits<int>::max(), std::numeric_limits<int>::max()), /*expected num items=*/ioIndices.size() * 3);
 	for (uint triangle_idx = 0; triangle_idx < ioIndices.size(); ++triangle_idx)
 	{
 		IndexedTriangle &triangle = ioIndices[triangle_idx];