@Ono-Sendai 2024-11-05
Here's a before optimisations vs. after optimisations profile trace: (from Tracy profiler)

![without jolt optimisations](https://github.com/user-attachments/assets/3bb3de50-60d6-429c-9534-abf5fcec69e1)

![with jolt optimisations](https://github.com/user-attachments/assets/88f6ac96-9c87-49f4-a1c6-f991fb05a77c)