@Ono-Sendai 2024-11-09
avoiding AABBTree node allocations could also be done with a custom allocator, might be easier.