@jrouwe 2024-11-09
> avoiding AABBTree node allocations could also be done with a custom allocator, might be easier.

Too late, it's integrated now.