@Ono-Sendai 2024-11-09
> > avoiding AABBTree node allocations could also be done with a custom allocator, might be easier.
> 
> Too late, it's integrated now.

haha ok.
I think the approach with a single vector is good for a single-threaded builder.  The approach I use for a multithreaded builder (an idea from Embree) is to allocate nodes in per-thread buffers of e.g. 256 nodes.   Then these are merged in a final pass in the main thread.