@jrouwe 2024-11-19
I implemented a hash map/set from scratch as I didn't want to reserve a separate value for an invalid entry. It is integrated in #1337 and in my tests it performs equally well as your hash map/set implementation. Let me know how this works for you.

That means that all parts of this PR are integrated now, so I'm going to close this PR! Thanks for pushing me to get rid of std::unordered_map/set, that had been on my list for a long time :)