@Ono-Sendai 2024-11-20
Ok sounds good, I'm sure your hash map/set are fine.